moved partition helper to util
