+++ q2_types/genome_data/_methods.py
@@ -16,7 +16,7 @@ import numpy as np
 import skbio
 from qiime2.util import duplicate
 
-from q2_types._util import _collate_helper
+from q2_types._util import _collate_helper, partition_helper
 from q2_types.feature_data import DNAIterator, DNAFASTAFormat
 from q2_types.genome_data import (
     SeedOrthologDirFmt, OrthologAnnotationDirFmt, LociDirectoryFormat,
@@ -170,58 +170,3 @@ def partition_loci(
     loci: LociDirectoryFormat, num_partitions: int = None
 ) -> LociDirectoryFormat:
     return partition_helper(loci, num_partitions)
-
-
-def partition_helper(dir_format, num_partitions: int = None):
-    """
-    This function splits the file dictionary of the given directory format into
-    a specified number of partitions. For each partition, a new instance of the
-    same directory format class is created and populated with the corresponding
-    files. If the values in the file dictionary are nested dictionaries,
-    subdirectories are created in the partition to preserve structure.
-
-    Parameters:
-        dir_format (qiime2.plugin.model.DirectoryFormat):
-            The directory format object containing files to partition.
-        num_partitions (int):
-            The number of partitions to split the files into. If None, the
-            function will partition by samples if applicable else by file.
-
-    Returns:
-        dict:
-            A dictionary mapping either partition indices or, if each
-            partition contains only a single sample/file, the corresponding
-            IDs, to new directory format instances containing the partitioned
-            files.
-    """
-    partitioned = {}
-    all = [{k: v} for k, v in dir_format.file_dict().items()]
-
-    num_partitions = _validate_num_partitions(
-        len(all), num_partitions, "sample"
-    )
-
-    arrays = np.array_split(all, num_partitions)
-
-    for i, samples in enumerate(arrays, 1):
-        result = dir_format.__class__()
-
-        for dict in samples:
-            if isinstance(next(iter(dict.values())), str):
-                for _id, fp in dict.items():
-                    duplicate(fp, result.path / os.path.basename(fp))
-            else:
-                for _id, feature_dict in dict.items():
-                    for fp in feature_dict.values():
-                        os.makedirs(result.path / _id, exist_ok=True)
-                        duplicate(
-                            fp,
-                            result.path / _id / os.path.basename(fp)
-                        )
-
-        if num_partitions == len(all):
-            partitioned[_id] = result
-        else:
-            partitioned[i] = result
-
-    return partitioned
