+++ q2_types/tests/data/genes/genes1.fa
@@ -1,0 +1,9 @@
+>gene1 some_description1
+GGCAGATTCCCCCTAGACCCGCCCGCACCATGGTCAGGCATGCCCCTCCTCATCGCTGGGCACAGCCCAGAGGGT
+ATAAACAGTGCTGGAGGC
+>gene2 some_description2
+CCACTGCACTCACCGCACCCGGCCAATTTTTGTGTTTTTAGTAGAGACTAAATACCATATAGTGAACACCTAAGA
+CGGGGGGCCTTGG
+>gene3 some_description3
+GCACCCGGCCAATTTTTGTGTTTTTAGTAGAGAAGATTCCCCCTAGACCCGCCCGCTATAGTGAACACCTAAGAA
+CTGGAGG
