+++ q2_types/tests/data/genes/genes2.fa
@@ -1,0 +1,6 @@
+>gene11 some_description11
+ATGGTCAGGCATGCCCCTCCTCATCGCTGGGCGGCAGATTCCCCCTAGACCCGCCCGCACCACAGCCCAGAGGGT
+ATAAACAGTGCTGGAGGC
+>gene12 some_description12
+AATACCATATAGTGAACACCTAACCACTGCACTCACCGCACCCGGCCAATTTTTGTGTTTTTAGTAGAGACTAGA
+CGGGGGGCCTTGG
