+++ q2_types/tests/data/genes_samples/sample2/genes2.fa
