+++ q2_types/tests/test_util.py
@@ -8,15 +8,18 @@
 import os
 import warnings
 from pathlib import Path
+from typing import TypeVar
 
+from q2_types.genome_data import GenesDirectoryFormat
 from q2_types.kraken2 import Kraken2OutputDirectoryFormat
-from qiime2.plugin import model
+from qiime2.plugin import model, DirectoryFormat
 from qiime2.plugin.testing import TestPluginBase
 
 from q2_types._util import (
     _validate_num_partitions, _validate_mag_ids, FileDictMixin,
-    _duplicate_with_warning
+    _duplicate_with_warning, partition_helper
 )
+DirFmt = TypeVar("DirFmt", bound=model.DirectoryFormat)
 
 
 class TestUtil(TestPluginBase):
@@ -190,3 +193,47 @@ class TestFileDictMixin(TestPluginBase):
 
         self.assertEqual(result_id, "sample_id")
         self.assertEqual(result_path, str(path.absolute()))
+
+    def test_partition_helper_genes_samples(self):
+        path = self.get_data_path("genes_samples")
+        genes = GenesDirectoryFormat(path=path, mode="r")
+        obs = partition_helper(dir_format=genes)
+        self.assertTrue(os.path.exists(
+            obs["sample1"].path / "sample1" / "genes1.fa")
+        )
+        self.assertTrue(os.path.exists(
+            obs["sample2"].path / "sample2" / "genes2.fa")
+        )
+
+    def test_partition_helper_genes(self):
+        path = self.get_data_path("genes")
+        genes = GenesDirectoryFormat(path=path, mode="r")
+        obs = partition_helper(dir_format=genes)
+        self.assertTrue(os.path.exists(
+            obs["genes1"].path / "genes1.fa")
+        )
+        self.assertTrue(os.path.exists(
+            obs["genes2"].path / "genes2.fa")
+        )
+
+    def test_partition_helper_genes_samples_1_partition(self):
+        path = self.get_data_path("genes_samples")
+        genes = GenesDirectoryFormat(path=path, mode="r")
+        obs = partition_helper(dir_format=genes, num_partitions=1)
+        self.assertTrue(os.path.exists(
+            obs[1].path / "sample1" / "genes1.fa")
+        )
+        self.assertTrue(os.path.exists(
+            obs[1].path / "sample2" / "genes2.fa")
+        )
+
+    def test_partition_helper_genes_1_partition(self):
+        path = self.get_data_path("genes")
+        genes = GenesDirectoryFormat(path=path, mode="r")
+        obs = partition_helper(dir_format=genes, num_partitions=1)
+        self.assertTrue(os.path.exists(
+            obs[1].path / "genes1.fa")
+        )
+        self.assertTrue(os.path.exists(
+            obs[1].path / "genes2.fa")
+        )
